fix(build): pin to Debian bookworm-based images instead of `trixie` #0000
One of the packages these images install is the `wkhtmltopdf`, which is
used by Onatal's legacy PDF renderer.

Since Debian Trixie was released on 2025 Aug, Docker Hub's PHP builds
are based on Trixie. While this works for all other packages, this
Debian version drops the then-obsolete `wkhtmltopdf` package:

 - [`wkhtmltopdf` on `bookworm`](https://packages.debian.org/search?suite=bookworm&keywords=wkhtmltopdf)
 - [`wkhtmltopdf` on `trixie` (does not exist)](https://packages.debian.org/search?suite=trixie&keywords=wkhtmltopdf)

To avoid build failures, this changes the PHP images to the ones tagged
Debian bookworm.