fix(reviewdog): fix installation and directly install to `/usr/local/bin` #0000
The `*-review` images include `reviewdog`, installed by `curl | sh`-ing
the `install.sh`.

Following a recent change[^1], the default installation directory has
changed.

This updates the `*-review` `Dockerfile`s to be compatible with the
`install.sh` changes by directly specifying the installation directory.

Previously, this required manually moving the executable, but from now
on, the installation directory is specified as a flag to the
`install.sh` script.

[^1]: https://github.com/reviewdog/reviewdog/commit/a547abecbdf6f6c69453be7c04a630ba939c27c0